src/modules/cars/useCases/listAvailableCars/ListAvailableCarsUseCase.ts
import { ICarRepository } from "@modules/cars/repositories/icar.repository";
import { inject, injectable } from "tsyringe";

interface IRequest {
  category_id?: string;
  brand?: string;
  name?: string;
}

@injectable()
export class ListAvailableCarsUseCase {
  constructor(
    @inject("CarRepository")
    private carsRepository: ICarRepository
  ) {}

  async execute({ category_id, name, brand }: IRequest) {
    const cars = await this.carsRepository.findAvailable(
      brand,
      category_id,
      name
    );
    return cars;
  }
}
